+++ src/libretro/api/input/pointer.py
@@ -8,6 +8,7 @@ RETRO_DEVICE_ID_POINTER_X = 0
 RETRO_DEVICE_ID_POINTER_Y = 1
 RETRO_DEVICE_ID_POINTER_PRESSED = 2
 RETRO_DEVICE_ID_POINTER_COUNT = 3
+RETRO_DEVICE_ID_POINTER_IS_OFFSCREEN = 15
 
 
 class DeviceIdPointer(IntEnum):
@@ -15,6 +16,7 @@ class DeviceIdPointer(IntEnum):
     Y = RETRO_DEVICE_ID_POINTER_Y
     PRESSED = RETRO_DEVICE_ID_POINTER_PRESSED
     COUNT = RETRO_DEVICE_ID_POINTER_COUNT
+    IS_OFFSCREEN = RETRO_DEVICE_ID_POINTER_IS_OFFSCREEN
 
     def __init__(self, value: int):
         self._type_ = "H"
@@ -25,6 +27,7 @@ class Pointer:
     x: int = 0
     y: int = 0
     pressed: bool = False
+    is_offscreen: bool = False
 
 
 @dataclass(frozen=True, slots=True)